+++ frontend/talentsearch/src/js/components/profile/ProfilePage/ProfilePage.tsx
@@ -25,7 +25,7 @@ import {
   disabilityLocalized,
 } from "@common/constants/localizedConstants";
 
-import { insertBetween } from "@common/helpers/util";
+import { insertBetween, notEmpty } from "@common/helpers/util";
 import TALENTSEARCH_APP_DIR from "../../../talentSearchConstants";
 import { useApplicantProfileRoutes } from "../../../applicantProfileRoutes";
 import {
@@ -35,6 +35,7 @@ import {
   GetMeQuery,
   GovEmployeeType,
 } from "../../../api/generated";
+import ExperienceSection from "../../applicantProfile/ExperienceSection";
 
 export interface ProfilePageProps {
   profileDataInput: User;
@@ -72,6 +73,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
     acceptedOperationalRequirements,
     wouldAcceptTemporary,
     poolCandidates,
+    experiences,
   } = profileDataInput;
 
   const intl = useIntl();
@@ -275,7 +277,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 })}
               </h2>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -317,7 +319,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -460,7 +462,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -681,7 +683,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -809,7 +811,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -895,7 +897,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -1004,7 +1006,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 </Link>
               </div>
               <div
-                data-h2-bg-color="b(gray)"
+                data-h2-bg-color="b(lightgray)"
                 data-h2-padding="b(all, m)"
                 data-h2-radius="b(s)"
               >
@@ -1062,7 +1064,7 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                 )}
               </div>
             </div>
-            <div id="skills-section">
+            <div id="skills-section" data-h2-padding="b(bottom, xxl)">
               <div style={{ display: "flex", alignItems: "baseline" }}>
                 <h2 data-h2-font-weight="b(600)" style={{ flex: "1 1 0%" }}>
                   <LightningBoltIcon style={{ width: "calc(1rem*2.25)" }} />
@@ -1087,7 +1089,49 @@ export const ProfileForm: React.FC<ProfilePageProps> = ({
                   })}
                 </Link>
               </div>
-              <p>Skill and experience details</p>
+              {console.log(experiences)}
+              {!experiences || experiences?.length === 0 ? (
+                <div
+                  data-h2-bg-color="b(lightgray)"
+                  data-h2-padding="b(all, m)"
+                  data-h2-radius="b(s)"
+                >
+                  <p>
+                    {intl.formatMessage({
+                      defaultMessage:
+                        "You haven't added any information here yet.",
+                      description:
+                        "Message that the user hasn't filled out the section yet",
+                    })}
+                  </p>
+                  <p>
+                    {intl.formatMessage(
+                      {
+                        defaultMessage:
+                          "There are <redText>required</redText> fields missing.",
+                        description:
+                          "Message that there are required fields missing. Please ignore things in <> tags.",
+                      },
+                      {
+                        redText,
+                      },
+                    )}{" "}
+                    <a href={paths.skillsAndExperiences()}>
+                      {intl.formatMessage({
+                        defaultMessage: "Click here to get started.",
+                        description:
+                          "Message to click on the words to begin something",
+                      })}
+                    </a>
+                  </p>
+                </div>
+              ) : (
+                <div data-h2-padding="b(all, m)" data-h2-radius="b(s)">
+                  <ExperienceSection
+                    experiences={experiences?.filter(notEmpty)}
+                  />
+                </div>
+              )}
             </div>
           </div>
         </div>
